Move help text into config
